Move Nav into Layout/ directory
